Frontend: Minor restructuring.
