+++ app/javascript/components/shared/stylings/Spinner.jsx
