+++ app/javascript/main.jsx
@@ -6,8 +6,8 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import App from './App';
 import Signup from './components/users/Signup';
-import SignIn from './components/sessions/SignIn';
-import AuthProvider from './components/sessions/AuthProvider';
+import SignIn from './components/users/SignIn';
+import AuthProvider from './contexts/auth/AuthProvider';
 import Room from './components/rooms/Room';
 import Rooms from './components/rooms/Rooms';
